add sliced block structures and sysfs for stats
Adds ouichefs_sliced_block which consists of 32 slices, the first for
metadata and the rest for file data. Necessary to access them is a new
superblock info field to store first partially filled block (new list).
This patch also adds a sysfs which shows stats of the partition. This
required adding new variables for tracking which also have to be written
to disk so mkfs had to be updated as well.

Co-authored-by: Jonathan Lyssy <jonathan.lyssy@rwth-aachen.de>