use new sliced blocks for files smaller than 128 bytes
This means mainly changing read and write. As a debugging help a
DUMP_BLOCK ioctl command was added that prints block content. The
created slice.c file contains every needed function for getting,
allocating and freeing slices. This includes a get_best_fit for the
slices as the next planned changes is the handling of files which are
stored in multiple contiguous slices.

While implementing, a concurency bug happened so that a bfree_lock was
introduced which is now used for accesses to the bfree_bitmap in
bitmap.h.

Co-authored-by: Jonathan Lyssy <jonathan.lyssy@rwth-aachen.de>