support multi-slice files with advanced allocation and relocation
Extend sliced blocks implementation by supporting files spanning
multiple slices. Allocation uses first-fit for blocks and best-fit for
the slice region in the block. To aid with this, the largest free gap of
a block is stored in the metadata slice. Whenever the file size changes
so that the number of necessary slices changes, the whole file data is
relocated. This could impact performance but the effect was not critical
and it helps fighting fragmentation. Fully functional, according to the
tests in the submission script.

Co-authored-by: Jonathan Lyssy <jonathan.lyssy@rwth-aachen.de>